+++ index.js
@@ -1,16 +1,4 @@
-import { createExpressApp } from "./server";
-import { logger } from "./logger";
-
-const port = parseInt(process.env.PORT, 10) || 3000;
-
-async function main() {
-  const server = await createExpressApp();
-  server.listen(port, err => {
-    if (err) {
-      throw err;
-    }
-    logger.info(`> Ready on http://localhost:${port}`);
-  });
-}
-
-main();
+require = require("esm")(module);
+const { createExpressApp } = require("./server/index.js");
+const app = createExpressApp();
+module.exports = app;
