--- now.js
@@ -1,4 +1,0 @@
-require = require("esm")(module);
-const { createExpressApp } = require("./server/index.js");
-const app = createExpressApp();
-module.exports = app;
