+++ now.json
@@ -1,5 +1,5 @@
 {
   "version": 2,
   "name": "goodWork",
-  "builds": [{ "src": "now.js", "use": "@now/node" }]
+  "builds": [{ "src": "index.js", "use": "@now/node" }]
 }
