+++ package.json
@@ -2,11 +2,11 @@
   "name": "v8-bears-team-06",
   "version": "0.0.1",
   "description": "Add-project-description-here | Voyage-8 | https://chingu.io/",
-  "main": "index.js",
+  "main": "main.js",
   "scripts": {
-    "start:dev": "nodemon index.js --exec babel-node",
+    "start:dev": "nodemon main.js --exec babel-node",
     "build": "next build ./client",
-    "start": "node -r esm index.js",
+    "start": "node -r esm main.js",
     "test": "jest",
     "test:watch": "jest --watchAll",
     "format": "prettier --write \"**/*.{js,jsx,json,md,yml,yaml,html,graphql}\"",
